src/BookManagerService/requirements.txt
flask==3.0.0
flask-consulate==0.2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2